Decouple attribute value adjustment/validation from ember.
Looking to make scenes generic (not depend on ember).
