+++ src/app/clusters/color-control-server/color-control-server.cpp
@@ -29,6 +29,7 @@
 #include <tracing/macros.h>
 
 #ifdef MATTER_DM_PLUGIN_SCENES_MANAGEMENT
+#include <app/clusters/scenes-server/CodegenAttributeValuePairValidator.h>
 #include <app/clusters/scenes-server/scenes-server.h>
 #endif
 
@@ -46,8 +47,8 @@ public:
     // be updated.
     static constexpr uint8_t kColorControlScenableAttributesCount = 9;
 
-    DefaultColorControlSceneHandler() = default;
-    ~DefaultColorControlSceneHandler() override {}
+    DefaultColorControlSceneHandler() : scenes::DefaultSceneHandlerImpl(scenes::CodegenAttributeValuePairValidator::Instance()) {}
+    ~DefaultColorControlSceneHandler() override = default;
 
     // Default function for ColorControl cluster, only checks if ColorControl is enabled on the endpoint
     bool SupportsCluster(EndpointId endpoint, ClusterId cluster) override
