+++ src/app/clusters/level-control/level-control.cpp
@@ -38,6 +38,7 @@
 #include <tracing/macros.h>
 
 #ifdef MATTER_DM_PLUGIN_SCENES_MANAGEMENT
+#include <app/clusters/scenes-server/CodegenAttributeValuePairValidator.h>
 #include <app/clusters/scenes-server/scenes-server.h>
 #endif // MATTER_DM_PLUGIN_SCENES_MANAGEMENT
 
@@ -135,8 +136,8 @@ public:
     // As per spec, 2 attributes are scenable in the level control cluster
     static constexpr uint8_t kLevelMaxScenableAttributes = 2;
 
-    DefaultLevelControlSceneHandler() = default;
-    ~DefaultLevelControlSceneHandler() override {}
+    DefaultLevelControlSceneHandler() : scenes::DefaultSceneHandlerImpl(scenes::CodegenAttributeValuePairValidator::Instance()) {}
+    ~DefaultLevelControlSceneHandler() = default;
 
     // Default function for LevelControl cluster, only checks if LevelControl is enabled on the endpoint
     bool SupportsCluster(EndpointId endpoint, ClusterId cluster) override
