+++ src/app/clusters/mode-select-server/mode-select-server.cpp
@@ -34,6 +34,7 @@
 #include <tracing/macros.h>
 
 #ifdef MATTER_DM_PLUGIN_SCENES_MANAGEMENT
+#include <app/clusters/scenes-server/CodegenAttributeValuePairValidator.h>
 #include <app/clusters/scenes-server/scenes-server.h>
 #endif // MATTER_DM_PLUGIN_SCENES_MANAGEMENT
 
@@ -155,8 +156,9 @@ public:
     // As per spec, 1 attribute is scenable in the mode select cluster
     static constexpr uint8_t kScenableAttributeCount = 1;
 
-    DefaultModeSelectSceneHandler() = default;
-    ~DefaultModeSelectSceneHandler() override {}
+    DefaultModeSelectSceneHandler() : scenes::DefaultSceneHandlerImpl(scenes::CodegenAttributeValuePairValidator::Instance()) {}
+
+    ~DefaultModeSelectSceneHandler() override = default;
 
     // Default function for mode select cluster, only checks if mode select is enabled on the endpoint
     bool SupportsCluster(EndpointId endpoint, ClusterId cluster) override
