+++ src/app/clusters/scenes-server/AttributeValuePairValidator.h
@@ -1,0 +1,40 @@
+/**
+ *    Copyright (c) 2025 Project CHIP Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+#pragma once
+
+#include <app/ConcreteClusterPath.h>
+#include <clusters/ScenesManagement/Structs.h>
+#include <lib/core/CHIPError.h>
+
+namespace chip::scenes {
+
+class AttributeValuePairValidator
+{
+public:
+    virtual ~AttributeValuePairValidator() = default;
+
+    using AttributeValuePairType = app::Clusters::ScenesManagement::Structs::AttributeValuePairStruct::Type;
+
+    /// Validate and adjust the given attribute as needed.
+    ///
+    /// It may:
+    ///   - check the given attribute is valid on the given cluster
+    ///   - adjust value ranges
+    ///
+    virtual CHIP_ERROR Validate(const app::ConcreteClusterPath & clusterPath, AttributeValuePairType & value) = 0;
+};
+
+} // namespace chip::scenes
