+++ src/app/clusters/scenes-server/CodegenAttributeValuePairValidator.h
@@ -1,0 +1,35 @@
+/**
+ *    Copyright (c) 2025 Project CHIP Authors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+#pragma once
+
+#include <app/clusters/scenes-server/AttributeValuePairValidator.h>
+
+namespace chip::scenes {
+
+class CodegenAttributeValuePairValidator : public AttributeValuePairValidator
+{
+public:
+    ~CodegenAttributeValuePairValidator() override = default;
+    CHIP_ERROR Validate(const app::ConcreteClusterPath & clusterPath, AttributeValuePairType & value) override;
+
+    static CodegenAttributeValuePairValidator & Instance()
+    {
+        static CodegenAttributeValuePairValidator sInstance;
+        return sInstance;
+    }
+};
+
+} // namespace chip::scenes
