+++ src/app/clusters/scenes-server/app_config_dependent_sources.cmake
@@ -16,6 +16,9 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
+    "${CLUSTER_DIR}/AttributeValuePairValidator.h"
+    "${CLUSTER_DIR}/CodegenAttributeValuePairValidator.cpp"
+    "${CLUSTER_DIR}/CodegenAttributeValuePairValidator.h"
     "${CLUSTER_DIR}/ExtensionFieldSets.h"
     "${CLUSTER_DIR}/ExtensionFieldSetsImpl.cpp"
     "${CLUSTER_DIR}/ExtensionFieldSetsImpl.h"
