+++ src/app/clusters/scenes-server/app_config_dependent_sources.gni
@@ -12,6 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 app_config_dependent_sources = [
+  "AttributeValuePairValidator.h",
+  "CodegenAttributeValuePairValidator.cpp",
+  "CodegenAttributeValuePairValidator.h",
   "ExtensionFieldSets.h",
   "ExtensionFieldSetsImpl.cpp",
   "ExtensionFieldSetsImpl.h",
